Close #3: Show state of toogle buttons
We use document.queryCommandState() and document.queryCommandEnabled()
to retrieve the current state, and update the toolbar buttons to
reflect that on mouseup and on keyup.

We do not make any browser specific adjustments, but rather rely on the
browser yielding sensible values, which is, unfortunately, not always
the case, as some browsers apparently report certain commands to be
enabled, even though they won't have any effect. However, we have not
been able to find it the other way round, so this appears still to be
better than to not report any states at all.