Close #49: Improve conversion success messages
Particularly with regard to import, but also for HTML/plain text
conversion, it appears to be appropriate to tell how many comments
actually have been converted.

We're being a bit sloppy with regard to plural forms here, because
`XH_numberPrefix()` is insufficient[1], and we don't want to roll our
own pluralization for this rarely used functionality.

[1] https://cmsimpleforum.com/viewtopic.php?f=29&t=13113