Fix #48: Wrong URLs may be generated
We have to add special treatment for the already escaped "page URL", so
we add a respective parameter to the constructor.  To make it easier to
get the current URL, we add a static `::getCurrent()` method.

We also rename some methods to more clearly mark them as getters, and
add a test case for the `Url` class.