Overhaul MainAdminController
We implement confirmations for all conversions, and redirect to the
overview on success.  To retain the success messages (which are
important in this case), we introduce `FlashMessage`.  To avoid any
misuse of the `action`s, we use a proper working `twocents_action`
which we fetch from the request.

Finally, we implement proper pluralization.