Fix #68: Unrelated XHRs are erroneously handled
The previous attempt to fix this is issue was bad, since

* we must not remove any parameters from the query, because that would
  break a browser refresh
* we still play God by possibly catching any XHR

Due to the first issue, we also can't add a special query parameter.
So the only (hopefully) viable solution is using a custom HTTP request
header; we call that `X-CMSimple-XH-Request` so it is generic enough to
be used elsewhere for CMSimple_XH, but special enough to not clash with
any other headers.