Avoid this in event handlers
If an event handler is set to a classic function expression, `this` is
bound to the `EvenTarget`.  However, in an arrow function, it is not.
Thus it seems prudent to avoid `this` in event handler functions.  We
can easily replace that with `Event.currentTarget`, although that
appears to be a misnomer.  The cleaner solution seems to be to bind to
an outer variable, but that may not necessarily work as expected.  Thus
we actually use `Event.currentTarget` for now.